+++ tools/ncurses/samples/attributes.c
@@ -2,25 +2,8 @@
 #include <string.h>
 #include <signal.h>
 #include <unistd.h>
+#include "color.h"
 
-enum color_enum {
-	C_GREEN,
-	C_RED,
-	C_CYAN,
-	C_WHITE,
-	C_MAGENTA,
-	C_BLUE,
-	C_YELLOW,
-	C_MAX,
-	C_UNKNOWN = C_MAX,
-};
-const char *color_names[] = {
-	[C_GREEN] = "GREEN",   [C_RED] = "RED",		[C_CYAN] = "CYAN",
-	[C_WHITE] = "WHITE",   [C_MAGENTA] = "MAGENTA", [C_BLUE] = "BLUE",
-	[C_YELLOW] = "YELLOW",
-};
-
-chtype colors[C_MAX] = { 0 };
 chtype attrs[] = {
 	A_NORMAL,     A_ATTRIBUTES, A_CHARTEXT, A_COLOR,      A_STANDOUT,
 	A_UNDERLINE,  A_REVERSE,    A_BLINK,	A_DIM,	      A_BOLD,
@@ -61,25 +44,6 @@ const char *attr_name(chtype a)
 			return __attr_names[i].name;
 	}
 	return "??";
-}
-
-void init_colors(void)
-{
-	if (has_colors()) {
-		start_color();
-#define SET_COLOR(num, fg)                                 \
-	init_pair(num + 1, (short)fg, (short)COLOR_BLACK); \
-	colors[num] |= (chtype)COLOR_PAIR(num + 1)
-
-		SET_COLOR(C_GREEN, COLOR_GREEN);
-		SET_COLOR(C_RED, COLOR_RED);
-		SET_COLOR(C_CYAN, COLOR_CYAN);
-		SET_COLOR(C_WHITE, COLOR_WHITE);
-		SET_COLOR(C_MAGENTA, COLOR_MAGENTA);
-		SET_COLOR(C_BLUE, COLOR_BLUE);
-		SET_COLOR(C_YELLOW, COLOR_YELLOW);
-#undef SET_COLOR
-	}
 }
 
 void print_attr(void)